fix: secure delete and put endpoints (#228)
* fix(poll): secure delete endpoint

* fix(poll): secure update endpoint

* feat: Forbidden message

* fix(participate): secure delete participation

* fix: allow admins to delete participations

* refactor(poll): move forbidden exception to controller

* refactor(participant): move forbidden exception to controller

* refactor(backend): imports

* fix: delete own participations

* refactor: remove unused poll id

* refactor: format imports

+++ apps/backend/src/poll/poll.controller.ts
@@ -7,6 +7,7 @@ import {
   Controller,
   DefaultValuePipe,
   Delete,
+  ForbiddenException,
   Get,
   Headers,
   Param,
@@ -71,8 +72,18 @@ export class PollController {
   }
 
   @Put(':id')
+  @UseGuards(OptionalAuthGuard)
   @NotFound()
-  async putPoll(@Param('id', ObjectIdPipe) id: Types.ObjectId, @Body() pollDto: PollDto): Promise<ReadPollDto | null> {
+  async putPoll(
+    @Headers('Participant-Token') token: string,
+    @Param('id', ObjectIdPipe) id: Types.ObjectId,
+    @Body() pollDto: PollDto,
+    @AuthUser() user?: UserToken,
+  ): Promise<ReadPollDto | null> {
+    const pollDoc = await this.pollService.find(id) ?? notFound(id);
+    if (!this.pollActionsService.isAdmin(pollDoc, token, user?.sub)) {
+      throw new ForbiddenException('You are not allowed to edit this poll');
+    }
     return this.pollActionsService.putPoll(id, pollDto);
   }
 
@@ -83,8 +94,17 @@ export class PollController {
   }
 
   @Delete(':id')
+  @UseGuards(OptionalAuthGuard)
   @NotFound()
-  async deletePoll(@Param('id', ObjectIdPipe) id: Types.ObjectId): Promise<ReadPollDto | null> {
+  async deletePoll(
+    @Headers('Participant-Token') token: string,
+    @Param('id', ObjectIdPipe) id: Types.ObjectId,
+    @AuthUser() user?: UserToken,
+  ): Promise<ReadPollDto | null> {
+    const pollDoc = await this.pollService.find(id) ?? notFound(id);
+    if (!this.pollActionsService.isAdmin(pollDoc, token, user?.sub)) {
+      throw new ForbiddenException('You are not allowed to delete this poll');
+    }
     return this.pollActionsService.deletePoll(id);
   }
